Crate measurements

Macros

implement_measurement!
This is a special macro that creates the code to implement operator and comparison overrides.

Structs

Length
The Length struct can be used to deal with lengths in a common way. Common metric and imperial units are supported.

Temperature
The Temperature struct can be used to deal with temperatures in a common way.

Volume
The Volume struct can be used to deal with volumes in a common way.

Weight
The Weight struct can be used to deal with weights in a common way.

Traits

Measurement
The Measurement trait and the implement_measurement! macro provides a common way for various measurements to be implemented.