Module measurements::measurement

The Measurement trait and the implement_measurement! macro provides a common way for various measurements to be implemented.

Example

#[macro_use]
use measurements::measurement::*;
 
struct Cubits {
    forearms: f64
}
 
impl Measurement for Cubits {
    fn get_base_units(&self) -> f64 {
        self.forearms
    }
     
    fn from_base_units(units: f64) -> Self {
        Cubits { forearms: units }
    }
}

// Invoke the macro to automatically implement Add, Sub, etc...
// implement_measurement! { Cubits } // TODO: Fix doc tests to make this work...

Traits

Measurement